8933: add feature spec for user batch file upload
