+++ spec/features/results/file_upload_spec.rb
@@ -89,17 +89,5 @@ feature "response form file upload", js: true do
     expect(video_node).to have_selector(".media-thumbnail img")
   end
 
-  def expect_preview(node)
-    expect(node).to have_selector(".dz-preview")
-    expect(node).to_not have_content("The uploaded file was not an accepted format.")
-  end
 
-  def expect_no_preview(node)
-    expect(node).to_not have_selector(".dz-preview")
-  end
-
-  def delete_file(node)
-    node.find(".delete").click
-    page.driver.browser.switch_to.alert.accept
-  end
 end
