+++ spec/support/contexts/dropzone.rb
@@ -25,6 +25,20 @@ shared_context "dropzone" do
     JS
   end
 
+  def expect_preview(node)
+    expect(node).to have_selector(".dz-preview")
+    expect(node).to_not have_content("The uploaded file was not an accepted format.")
+  end
+
+  def expect_no_preview(node)
+    expect(node).to_not have_selector(".dz-preview")
+  end
+
+  def delete_file(node)
+    node.find(".delete").click
+    page.driver.browser.switch_to.alert.accept
+  end
+
   private
 
   def wait_for_dropzone_upload
